Move bundle fetching into IndexService
